% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_randomization.R
\name{randomize_choice_order}
\alias{randomize_choice_order}
\title{Randomize Choice Order}
\usage{
randomize_choice_order(
  choice_sheet,
  freeze_top = NULL,
  freeze_bottom = NULL,
  pth = NULL,
  output = TRUE
)
}
\arguments{
\item{choice_sheet}{A (path to a) ODK xlsx cohice sheet that has a set of programmed choices}

\item{freeze_top}{Optional - Number of rows at the top of the choice list that should stay frozen. Defaults to NULL}

\item{freeze_bottom}{Optional - Number of rows at the bottom of the choice list that should stay froze. Defaults to NULL. If both `freeze_top` and `freeze_bottom` are NULL, items with a value less than 90 will be randomized.}

\item{pth}{Optional - A file path to where the output should be saved. Defaults to the current working directory}

\item{output}{Optional - If TRUE will return the xlsx form as a data frame to R. If FALSE the only output is the xlsx file. Defaults to TRUE}
}
\value{
A ODK-style choice sheet as a data frame and an .xlsx sheet
}
\description{
Given a single set of choices already programmed on an xlsx sheet in ODK format,
this function will create an xlsx sheet containing all the possible permutations
of the choice order with any choice of value 90 or higher frozen at the end of
the order by default. It is also possible to freeze choice orders at the start
of the list, or at the start and end.

It is then possible to directly copy and paste the output
into the XLSForm, but it is recommended to upload this xlsx form to KoBoToolbox
and use either `select_one_from_file` or `select_multiple_from_file` question
type and reference the output file.

**These questions will still require a calculate type question in the survey sheet to determine which set of choices to display.**
The user can copy the calculation printed out by the function in a warning message and paste it
into the **calculation** column for a **calculate** type question.
}
\details{
The user must provide the **number of rows** they want to freeze at the top and/or
bottom of the choice list. For example, if there are seven options but the user
wants the first item to always appear first and the last two items to always appear
last. The `freeze_top` parameter should then be set to `1` and the `freeze_bottom`
parameter should be set to `2`.
}
